refactor: split mocks and rename test files
