+++ internal/cmd/local/local/mock_test.go
@@ -1,0 +1,98 @@
+package local
+
+import (
+	"context"
+	"net/http"
+
+	"github.com/airbytehq/abctl/internal/cmd/local/helm"
+	"github.com/airbytehq/abctl/internal/telemetry"
+	"github.com/google/uuid"
+	helmclient "github.com/mittwald/go-helm-client"
+	"helm.sh/helm/v3/pkg/action"
+	"helm.sh/helm/v3/pkg/chart"
+	"helm.sh/helm/v3/pkg/release"
+	"helm.sh/helm/v3/pkg/repo"
+)
+
+// ---
+// only mocks below here
+// ---
+var _ helm.Client = (*mockHelmClient)(nil)
+
+type mockHelmClient struct {
+	addOrUpdateChartRepo   func(entry repo.Entry) error
+	getChart               func(string, *action.ChartPathOptions) (*chart.Chart, string, error)
+	getRelease             func(name string) (*release.Release, error)
+	installOrUpgradeChart  func(ctx context.Context, spec *helmclient.ChartSpec, opts *helmclient.GenericHelmOptions) (*release.Release, error)
+	uninstallReleaseByName func(s string) error
+}
+
+func (m *mockHelmClient) AddOrUpdateChartRepo(entry repo.Entry) error {
+	return m.addOrUpdateChartRepo(entry)
+}
+
+func (m *mockHelmClient) GetChart(s string, options *action.ChartPathOptions) (*chart.Chart, string, error) {
+	return m.getChart(s, options)
+}
+
+func (m *mockHelmClient) GetRelease(name string) (*release.Release, error) {
+	return m.getRelease(name)
+}
+
+func (m *mockHelmClient) InstallOrUpgradeChart(ctx context.Context, spec *helmclient.ChartSpec, opts *helmclient.GenericHelmOptions) (*release.Release, error) {
+	return m.installOrUpgradeChart(ctx, spec, opts)
+}
+
+func (m *mockHelmClient) UninstallReleaseByName(s string) error {
+	return m.uninstallReleaseByName(s)
+}
+
+var _ telemetry.Client = (*mockTelemetryClient)(nil)
+
+type mockTelemetryClient struct {
+	start   func(context.Context, telemetry.EventType) error
+	success func(context.Context, telemetry.EventType) error
+	failure func(context.Context, telemetry.EventType, error) error
+	attr    func(key, val string)
+	user    func() uuid.UUID
+	wrap    func(context.Context, telemetry.EventType, func() error) error
+}
+
+func (m *mockTelemetryClient) Start(ctx context.Context, eventType telemetry.EventType) error {
+	return m.start(ctx, eventType)
+}
+
+func (m *mockTelemetryClient) Success(ctx context.Context, eventType telemetry.EventType) error {
+	return m.success(ctx, eventType)
+}
+
+func (m *mockTelemetryClient) Failure(ctx context.Context, eventType telemetry.EventType, err error) error {
+	return m.failure(ctx, eventType, err)
+}
+
+func (m *mockTelemetryClient) Attr(key, val string) {
+	if m.attr != nil {
+		m.attr(key, val)
+	}
+}
+
+func (m *mockTelemetryClient) User() uuid.UUID {
+	if m.user == nil {
+		return uuid.Nil
+	}
+	return m.user()
+}
+
+func (m *mockTelemetryClient) Wrap(ctx context.Context, et telemetry.EventType, f func() error) error {
+	return m.wrap(ctx, et, f)
+}
+
+var _ HTTPClient = (*mockHTTP)(nil)
+
+type mockHTTP struct {
+	do func(req *http.Request) (*http.Response, error)
+}
+
+func (m *mockHTTP) Do(req *http.Request) (*http.Response, error) {
+	return m.do(req)
+}
